Refactored routes to MVC methodology
